go/src/samsaradev.io/infra/dataplatform/dagster/projects/ml/requirements.txt
dagster==1.9.1
dagster-webserver==1.9.1
dagster-aws==0.25.1
dagster-databricks==0.25.1
dagster-pyspark==0.25.1
dagster-slack==0.25.1
dagster-postgres==0.25.1
numpy==2.2.6
pandas==2.3.2
scikit-learn==1.7.0
scipy==1.14.1
h3==4.3.1
shapely==2.1.2
datadog==0.42.0
slackclient==2.8.0
databricks-sql-connector==2.9.3
requests==2.29.0
pyproj==3.7.1
pdpyras==5.2.0
jsonschema==4.17.3
